# experiment with streaming JSON serialization