@Geal 2022-01-25
## main 2fe563d
279MB max resident memory usage

Summary:                                                         
  Total:        81.6236 secs                                     
  Slowest:      3.3268 secs                                      
  Fastest:      0.0052 secs                                      
  Average:      0.2601 secs                                      
  Requests/sec: 3675.4057                                        
                                                                 
  Total data:   3576900000 bytes                                 
  Size/request: 11923 bytes                                      
                                                                 
Response time histogram:                                         
  0.005 [1]     |                                                
  0.337 [220261]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.669 [26984] |■■■■■                                           
  1.002 [41544] |■■■■■■■■                                        
  1.334 [5006]  |■                                               
  1.666 [4912]  |■                                               
  1.998 [822]   |                                                
  2.330 [386]   |                                                
  2.662 [53]    |                                                
  2.995 [25]    |                                                
  3.327 [6]     |                                                
                                                                 
                                                                 
Latency distribution:                                            
  10% in 0.0482 secs                                             
  25% in 0.0575 secs                                             
  50% in 0.0706 secs                                             
  75% in 0.5126 secs                                             
  90% in 0.7731 secs                                             
  95% in 0.9039 secs                                             
  99% in 1.4855 secs                                             


## PR fceed7f (serializing in spawn_blocking)
221MB res

Summary:                                                          
  Total:        78.9947 secs                                      
  Slowest:      2.2936 secs                                       
  Fastest:      0.0054 secs                                       
  Average:      0.2543 secs                                       
  Requests/sec: 3797.7223                                         
                                                                  
                                                                  
Response time histogram:                                          
  0.005 [1]     |                                                 
  0.234 [196481]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■ 
  0.463 [9183]  |■■                                               
  0.692 [77281] |■■■■■■■■■■■■■■■■                                 
  0.921 [9539]  |■■                                               
  1.149 [4442]  |■                                                
  1.378 [2638]  |■                                                
  1.607 [201]   |                                                 
  1.836 [204]   |                                                 
  2.065 [20]    |                                                 
  2.294 [10]    |                                                 
                                                                  
                                                                  
Latency distribution:                                             
  10% in 0.0450 secs                                              
  25% in 0.0549 secs                                              
  50% in 0.0705 secs                                              
  75% in 0.5338 secs                                              
  90% in 0.6480 secs                                              
  95% in 0.7034 secs                                              
  99% in 1.1520 secs                                              


# PR f511dc9 (using tokio's DuplexStream and ReaderStream)

254MB res

Summary:                                                         
  Total:        92.8599 secs                                     
  Slowest:      3.8801 secs                                      
  Fastest:      0.0042 secs                                      
  Average:      0.2965 secs                                      
  Requests/sec: 3230.6749                                        
                                                                 
                                                                 
Response time histogram:                                         
  0.004 [1]     |                                                
  0.392 [209865]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.779 [53784] |■■■■■■■■■■                                      
  1.167 [24612] |■■■■■                                           
  1.555 [8765]  |■■                                              
  1.942 [2187]  |                                                
  2.330 [624]   |                                                
  2.717 [127]   |                                                
  3.105 [30]    |                                                
  3.492 [3]     |                                                
  3.880 [2]     |                                                
                                                                 
                                                                 
Latency distribution:                                            
  10% in 0.0543 secs                                             
  25% in 0.0648 secs                                             
  50% in 0.0799 secs                                             
  75% in 0.5947 secs                                             
  90% in 0.8128 secs                                             
  95% in 0.9787 secs                                             
  99% in 1.5526 secs                                             

I suspect that last attempt reallocates a lot instead of using fixed size buffers. It also does not benefit from serde_json_bytes: it would be possible to generate a Stream of Bytes from it without copying too much. That's the next experiment.