@Geal 2022-01-26
alright, I've done enough exploration on this. To sum up:
- current main serializes the response to a Vec (possibly reallocating multiple times), then sends it at once
- b094534 and 760e11e serialize in a blocking task, using a writer that allocates a big vector at once, and sends it through a channel once it's full. We build a hyper `Body` from a stream of buffers. **slightly faster than main, but more importantly: higher p50 but lower p99 (half). latency is much more stable**
- f511dc9 is the async version, serializing manually (without serde) in async. A bit slower, maybe because the serialization function is async and recursive, so allocating lots of futures?
- 9540b5f and 4ca35ce are dead ends. The idea was that the JSON object contains a Bytes for each string, so instead of copying into a buffer we can send those. Unfortunately hyper does not handle well a large number of small buffers, so we get -83% perf
- a53a711 (async) and eb10e5e (sync) refine that idea by serializing to a buffer, but when we get a string larger than the buffer, we flush the current buffer then send the string's Bytes right after